Login and Registration

+++ routes/web.php
@@ -22,24 +22,36 @@ Route::get('/', [CourseController::class, 'index'] );
 
 
 //Store Listing Data
-Route::post('/courses', [CourseController::class, 'store']);
+Route::post('/courses', [CourseController::class, 'store'])->middleware('auth');
 
 //Show Edit Form
-Route::get('/courses/{course}/edit',[CourseController::class, 'edit']);
+Route::get('/courses/{course}/edit',[CourseController::class, 'edit'])->middleware('auth');
 
 //Update Course
-Route::put('/courses/{course}', [CourseController::class, 'update']);
+Route::put('/courses/{course}', [CourseController::class, 'update'])->middleware('auth');
 
 //Delete Course
-Route::delete('/courses/{course}', [CourseController::class, 'destroy']);
+Route::delete('/courses/{course}', [CourseController::class, 'destroy'])->middleware('auth');
 
 
 //Show Create Form
-Route::get('/courses/create', [CourseController::class, 'create']);
+Route::get('/courses/create', [CourseController::class, 'create'])->middleware('auth');
 
-//Single Listing
+//Single Course
 Route::get('/courses/{course}',[CourseController::class, 'show']);
 
 
 //Show Register Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+//Create New User
+Route::post('/users', [UserController::class, 'store']);
+
+//Log User Out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show Login Form
+Route::get('/login' , [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Login User
+Route::post('users/authenticate', [UserController::class, 'authenticate']);
